requirements.txt
discord.py
aiosqlite
asyncio
schedule
datetime